PathFinding ALPHA
A* added.  Copied straight over from the workshop file so needs some
tidying but it does work. (Test this by putting a breakpoint where it
says done = true; at the end of the BuildPath method. Going to work on
adding a visualisation of the path to show this tomorrow.  Haven't
tested with other units on the map of different allegiances yet either.